+++ server.js
@@ -1,37 +1,27 @@
-const debug = require('debug')('urungi:server');
+const config = require('config');
 const cluster = require('cluster');
-const migrate = require('migrate-mongo');
 
 if (cluster.isMaster) {
-    migrate.database.connect().then(db => {
-        return migrate.status(db).then(statuses => {
-            if (statuses.some(s => s.appliedAt === 'PENDING')) {
-                throw new Error('Some migrations are pending. Run `npx migrate-mongo up` before starting the server');
-            }
-        });
-    }).then(function () {
-        var numCPUs = require('os').cpus().length;
+    const workers = config.get('workers');
 
-        // Fork workers.
-        for (var i = 0; i < numCPUs; i++) {
-            cluster.fork();
-        }
+    // Fork workers.
+    for (var i = 0; i < workers; i++) {
+        cluster.fork();
+    }
 
-        cluster.on('exit', function (deadWorker, code, signal) {
-            cluster.fork();
-        });
-    }, function (error) {
-        console.error('Error: ' + error.message);
-        process.exitCode = 1;
+    process.on('SIGTERM', terminate);
+    process.on('SIGINT', terminate);
+
+    cluster.on('exit', function () {
+        cluster.fork();
     });
 } else {
-    const config = require('config');
-    const app = require('./server/app');
+    require('./bin/www');
+}
 
-    var ipaddr = process.env.IP || config.get('ip');
-    var port = process.env.PORT || config.get('port');
-
-    const server = app.listen(port, ipaddr);
-    server.setTimeout(0);
-    debug('Server running at http://' + ipaddr + ':' + port + '/');
+function terminate () {
+    for (const id in cluster.workers) {
+        cluster.workers[id].kill();
+    }
+    process.exit(0);
 }